Refactor project to use Poetry
