--- environment.yml
@@ -1,13 +1,0 @@
-name: glint-mask-tools
-channels:
-  - conda-forge
-dependencies:
-  - python=3.8
-  - pillow
-  - numpy
-  - scipy
-  - fire
-  - tqdm
-  - pyqt
-  - loguru
-  - rasterio
