--- requirements.txt
@@ -1,8 +1,0 @@
-fire~=0.3.1
-numpy==1.22.0
-pillow~=8.3.1
-scipy~=1.5.2
-tqdm~=4.49.0
-pyqt5~=5.12.3
-loguru~=0.5.0
-rasterio~=1.3.4
