+++ src/glint_mask_tools/image_loaders.py
