+++ src/resources/buffer_ctrl.ui
