+++ src/resources/gmt.ico
