+++ src/resources/gui.ui
@@ -383,19 +383,19 @@
   <customwidget>
    <class>ThresholdCtrl</class>
    <extends>QWidget</extends>
-   <header>widgets/threshold_ctrl.h</header>
+   <header>src/widgets/threshold_ctrl.h</header>
    <container>1</container>
   </customwidget>
   <customwidget>
    <class>DirectoryPath</class>
    <extends>QWidget</extends>
-   <header>widgets/directory_path.h</header>
+   <header>src/widgets/directory_path.h</header>
    <container>1</container>
   </customwidget>
   <customwidget>
    <class>BufferCtrl</class>
    <extends>QWidget</extends>
-   <header>widgets/buffer_ctrl.h</header>
+   <header>src/widgets/buffer_ctrl.h</header>
    <container>1</container>
   </customwidget>
  </customwidgets>
