+++ src/widgets/directory_path.py
@@ -9,8 +9,8 @@ from os import path
 from PyQt5 import QtWidgets, uic
 from PyQt5.QtWidgets import QFileDialog
 
-bundle_dir = getattr(sys, '_MEIPASS', path.abspath(path.dirname(__file__)))
-UI_PATH = path.abspath(path.join(bundle_dir, 'directory_path.ui'))
+bundle_dir = getattr(sys, '_MEIPASS', path.abspath(path.dirname(path.dirname(__file__))))
+UI_PATH = path.abspath(path.join(bundle_dir, 'resources/directory_path.ui'))
 
 
 class DirectoryPath(QtWidgets.QWidget):
