+++ src/widgets/threshold_ctrl.py
@@ -9,8 +9,8 @@ from os import path
 
 from PyQt5 import QtWidgets, uic
 
-bundle_dir = getattr(sys, '_MEIPASS', path.abspath(path.dirname(__file__)))
-UI_PATH = path.abspath(path.join(bundle_dir, 'threshold_ctrl.ui'))
+bundle_dir = getattr(sys, '_MEIPASS', path.abspath(path.dirname(path.dirname(__file__))))
+UI_PATH = path.abspath(path.join(bundle_dir, 'resources/threshold_ctrl.ui'))
 
 
 class ThresholdCtrl(QtWidgets.QWidget):
